CRUD_With_FAISS_FastAPI/requirements.txt
faiss-cpu
fastapi
uvicorn
transformers
sentence-transformers
torch